Split bilevel into new tutorial.
